% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drpca.R
\name{Drpca}
\alias{Drpca}
\title{Distributed random PCA}
\usage{
Drpca(data,K, nk,d)
}
\arguments{
\item{data}{is  sparse random projection matrix}

\item{K}{is  the number of distributed nodes.}

\item{nk}{is the size of subsets.}

\item{d}{is the dimension number.
n is the sample size.
p the number of variables.}
}
\value{
MSEXrp, MSEvrp, kSopt, kxopt
}
\description{
Distributed random PCA
}
\examples{
K=20; nk=50; nr=50; p=8;d=5; n=K*nk;
data=matrix(c(rnorm((n-nr)*p,0,1),rpois(nr*p,100)),ncol=p)
}
