% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPC.R
\name{FPC}
\alias{FPC}
\title{FPC}
\usage{
FPC(data,K,nk)
}
\arguments{
\item{data}{is a data set}

\item{K}{is an index subset/sub-vectoris}

\item{nk}{is an index subset/sub-vectoris for each block}
}
\value{
MSEv1,MSEv2,MSEvopt,MSESopt1,MSESopt2,MSESopt,MSEShat,MSESba,MSESw
}
\description{
FPC
}
\examples{
library(matrixcalc)
K=20; nk=500; p=8; n=10000;m=50
data=matrix(c(rnorm((n-m)*p,0,1),rpois(m*p,100)),ncol=p)
FPC(data=data,K=K,nk=nk)
}
